updatepassword folders and edge cases
